+++ Source/Forms/EventCommandEditor.vb
@@ -382,7 +382,7 @@ Public Class EventCommandEditor
 
         Dim allCriteria As New List(Of Criteria)
 
-        For Each m In Macro.GetMacros(False, False)
+        For Each m In Macro.GetMacros(False, False, True)
             Dim c = Criteria.Create(m.Type)
             c.Name = m.FriendlyName
             c.Description = m.Description