+++ Source/Forms/MacrosForm.vb
@@ -287,7 +287,7 @@ Public Class MacrosForm
 
         Dim macros As New StringPairList
 
-        For Each mac In Macro.GetMacros(True, True)
+        For Each mac In Macro.GetMacros(True, True, True)
             macros.Add(mac.Name, mac.Description)
         Next