+++ Source/Forms/MuxerForm.vb
@@ -1321,7 +1321,7 @@ Public Class MuxerForm
             "A muxer merges different video, audio and subtitle files " +
             "into a single container file which is the actual output file.")
         form.Doc.WriteTips(TipProvider.GetTips, SimpleUI.ActivePage.TipProvider.GetTips)
-        form.Doc.WriteTable("Macros", Macro.GetTips(False, True))
+        form.Doc.WriteTable("Macros", Macro.GetTips(False, True, True))
         form.Show()
     End Sub