+++ Source/General/Proc.vb
@@ -405,7 +405,7 @@ Public Class Proc
 
         Dim keys = dic.Keys.OfType(Of String).Select(Function(key) key.ToLowerInvariant)
 
-        For Each mac In Macro.GetMacros(False, False)
+        For Each mac In Macro.GetMacros(False, False, True)
             Dim name = mac.Name.Trim("%"c)
 
             If Not keys.Contains(name) Then